organize cdk
